Make the XML API "UTF-8 clean"
Grails' render function for XML is idiotic wrt UTF-8 I think,
it doesn't seem to set the encoding for the underlying
Groovy StreamingMarkupWriter, so this defaults to US_ASCII
and the result is chaos: non-ascii letters all become escaped
as hex html entities.

For german this doesn't seem to be such a big deal, the
problem is only with a few characters like those with umlauts
and it can be worked around in the clients with some ugly
string replace hacks (see the online android client for
an example).

For Greek we would have to replace every single letter
which would be waaaaaaaaaaaaaaaay too annoying. So let's
fix the problem in the right place instead and have the
API send UTF-8 characters to the client.

I found out about this issue by porting the online version
of the android client myself (my first public "Android App" :P)
This shows the value of "dogfooding" in finding such API problems.