Include hacked version of spring web jar
Spring Framework has this idiotic behavior where it wants to escape
Greek characters to HTML entities no matter what. Well, there is
really no reason for this, but they don't seem to care to fix it
anytime soon either (there is a bug in their tracker for this with
no responses).

So, let's include our own hacked version of spring where we have
removed this 'feature' so that we send pure, clean UTF-8 Greek
to the browser the way god intended.